# sdks/sdkwork-knowledgebase-backend-sdk/openapi/knowledgebase-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Knowledgebase Backend API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/backend/v3/api/knowledge/okf/exports/{exportId}": {
      "get": {
        "operationId": "okf.bundle.export.retrieve",
        "tags": [
          "knowledge"
        ],
        "parameters": [
          {
            "name": "exportId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "uint64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/KnowledgeOkfBundleFile"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-knowledgebase",
        "x-sdkwork-api-authority": "sdkwork-knowledgebase-backend-api",
        "description": "Retrieve an OKF bundle export",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-knowledgebase-backend-api",
        "summary": "Retrieve an OKF bundle export",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "knowledge.admin",
        "x-sdkwork-audit-event": "knowledge.backend.admin_operation",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization"
      }
    },
...
    "/backend/v3/api/knowledge/indexes/{indexId}": {
      "get": {
        "operationId": "indexes.retrieve",
        "tags": [
          "knowledge"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/KnowledgeIndex"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-knowledgebase",
        "x-sdkwork-api-authority": "sdkwork-knowledgebase-backend-api",
        "parameters": [
          {
            "name": "indexId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "x-sdkwork-int64-string": true
            }
          }
        ],
        "description": "Retrieve a knowledge index",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-knowledgebase-backend-api",
        "summary": "Retrieve a knowledge index",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "knowledge.admin",
        "x-sdkwork-audit-event": "knowledge.backend.admin_operation",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization"
      }
    },
...
    "/backend/v3/api/knowledge/retrieval_profiles/{profileId}": {
      "get": {
        "operationId": "retrievalProfiles.retrieve",
        "tags": [
          "knowledge"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/KnowledgeRetrievalProfile"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-knowledgebase",
        "x-sdkwork-api-authority": "sdkwork-knowledgebase-backend-api",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "x-sdkwork-int64-string": true
            }
          }
        ],
        "description": "Retrieve a retrieval profile",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-knowledgebase-backend-api",
        "summary": "Retrieve a retrieval profile",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "knowledge.admin",
        "x-sdkwork-audit-event": "knowledge.backend.admin_operation",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization"
      },
      "patch": {
        "operationId": "retrievalProfiles.update",
        "tags": [
          "knowledge"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/KnowledgeRetrievalProfile"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-knowledgebase",
        "x-sdkwork-api-authority": "sdkwork-knowledgebase-backend-api",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeRetrievalProfileRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]+$",
              "x-sdkwork-int64-string": true
            }
          }
        ],
        "description": "Update a retrieval profile",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-source-route-crate": "sdkwork-routes-knowledgebase-backend-api",
        "summary": "Update a retrieval profile",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "knowledge.admin",
        "x-sdkwork-audit-event": "knowledge.backend.admin_operation",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "KnowledgeOkfBundleFile": {
        "type": "object",
        "required": [
          "id",
          "spaceId",
          "logicalPath",
          "entryType",
          "artifactRole",
          "driveBucket",
          "driveObjectKey"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint64"
          },
          "spaceId": {
            "type": "integer",
            "format": "uint64"
          },
          "logicalPath": {
            "type": "string"
          },
          "entryType": {
            "$ref": "#/components/schemas/OkfBundleFileKind"
          },
          "artifactRole": {
            "type": "string"
          },
          "driveBucket": {
            "type": "string"
          },
          "driveObjectKey": {
            "type": "string"
          },
          "checksumSha256Hex": {
            "type": [
              "string",
              "null"
            ]
          },
          "stagedImportRoot": {
            "type": "string"
          },
          "importId": {
            "type": "string"
          }
        }
      },
      "OkfBundleFileKind": {
        "type": "string",
        "enum": [
          "bundle_profile",
          "bundle_index",
          "bundle_log",
          "concept_revision",
          "graph_export",
          "context_pack",
          "output_export"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "traceId": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "KnowledgeIndex": {
        "type": "object",
        "required": [
          "indexId",
          "tenantId",
          "spaceId",
          "indexKind",
          "status"
        ],
        "properties": {
          "indexId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true
          },
          "tenantId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true
          },
          "spaceId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true
          },
          "indexKind": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "KnowledgeRetrievalProfileRequest": {
        "type": "object",
        "required": [
          "name",
          "strategy",
          "topK",
          "rerankEnabled",
          "contextBudgetTokens",
          "status"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "strategy": {
            "type": "string"
          },
          "topK": {
            "type": "integer",
            "format": "int32"
          },
          "minScore": {
            "anyOf": [
              {
                "type": "number",
                "format": "double"
              },
              {
                "type": "null"
              }
            ]
          },
          "rerankEnabled": {
            "type": "boolean"
          },
          "contextBudgetTokens": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "KnowledgeRetrievalProfile": {
        "type": "object",
        "required": [
          "retrievalProfileId",
          "tenantId",
          "name",
          "strategy",
          "topK",
          "rerankEnabled",
          "contextBudgetTokens",
          "status"
        ],
        "properties": {
          "retrievalProfileId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true
          },
          "tenantId": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true
          },
          "name": {
            "type": "string"
          },
          "strategy": {
            "type": "string"
          },
          "topK": {
            "type": "integer",
            "format": "int32"
          },
          "minScore": {
            "anyOf": [
              {
                "type": "number",
                "format": "double"
              },
              {
                "type": "null"
              }
            ]
          },
          "rerankEnabled": {
            "type": "boolean"
          },
          "contextBudgetTokens": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      }
    }
  },
...
  "servers": [
    {
      "url": "/backend/v3/api"
    }
  ]
}